# docs/openapi.yaml
openapi: 3.1.0
info:
  title: mKey daemon API
  version: 1.0.0
  description: |
    HTTP API демона mKey (SPEC §9, модуль internal/api).

    Транспорты:
    - Unix-сокет `$XDG_RUNTIME_DIR/mkey/mkey.sock` (права 0600) — для CLI, без токена;
    - `http://127.0.0.1:17420` — для веб-интерфейса: заголовок `Host` только 127.0.0.1/localhost,
      для изменяющих запросов `Origin` — только страницы mKey, токен из `$XDG_RUNTIME_DIR/mkey/token`
      в заголовке `Authorization: Bearer …` или cookie `mkey_session` (ставится при входе по `/?t=<токен>`).

    Сообщения ошибок переводятся на язык из `Accept-Language` (ru, en).
servers:
  - url: http://127.0.0.1:17420
security:
  - bearer: []
  - cookie: []
paths:
...
  /api/v1/send:
    post:
      summary: Выполнить макрос (ответ — после завершения; разрыв соединения прерывает макрос)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sequence]
              additionalProperties: false
              properties:
                sequence: { type: string, example: '^{Shift}{"Привет, Вера!"}~{Shift}{Enter}' }
                dry_run: { type: boolean, description: "Только проверить, ничего не нажимая; в ответе — план шагов" }
      responses:
        "200":
          description: Макрос выполнен (или проверен)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  duration_ms: { type: integer }
                  steps: { type: array, items: { type: object }, description: "Только при dry_run" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/dsl/parse:
    post:
      summary: Разобрать макрос в дерево (для конструктора блоков)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TextRequest" }
      responses:
        "200":
          description: Дерево разбора
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes: { type: array, items: { $ref: "#/components/schemas/Node" } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/dsl/format:
    post:
      summary: Привести макрос к каноническому виду
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TextRequest" }
      responses:
        "200":
          description: Канонический текст
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string, example: "^{Shift}{A}[1s]" }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/devices:
    get:
      summary: "Физические устройства ввода; auto_ids — авто-ID устройств (UnKey…) по пути, если работает инспектор"
      responses:
        "200":
          description: Устройства и доступ к ним
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices: { type: array, items: { type: object } }
                  status: { $ref: "#/components/schemas/InputStatus" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/devices/inspect:
    get:
      summary: "Подробности устройства (?ref= путь, eventN, постоянное имя by-id/by-path или часть названия): {device: DeviceDetails} — постоянные имена, шина, авто-ID устройства (auto_id) и кнопок (label: UnKey001), кнопки с именами mKey и ядра, оси с диапазонами, колёса, переключатели, индикаторы, отдача, свойства"
      parameters: [ { name: ref, in: query, required: true, schema: { type: string } } ]
      responses:
        "200": { description: "{device}" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: "api.device_ambiguous: подходит несколько устройств; details.candidates — [{path, name}]" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/devices/rename:
    post:
      summary: "Дать имя устройству или его кнопке (FR-DEV-3): {device, control, name}; control пусто — устройство; name пусто — убрать имя. Ошибка имени — 400 api.rename_chars|key|auto|taken|unknown|ambiguous"
      responses: { "200": { description: "{ok}" }, "400": { $ref: "#/components/responses/Error" }, "503": { $ref: "#/components/responses/Error" } }
  /api/v1/devices/virtual:
    get:
      summary: "Виртуальные устройства включённых проектов (ADR-0028): {devices: [{name, template, project, system_name, node?, error?}], templates: [...], template_info: [{id, buttons: [South…], axes: [LX…]}]} — состав шаблонов для мастера «второй геймпад»"
      responses: { "200": { description: "{devices, templates, template_info}" }, "503": { $ref: "#/components/responses/Error" } }
  /api/v1/settings/devices:
    get:
      summary: "Настройки устройств: {auto_ids} — каким устройствам давать авто-ID (smart, all, unusual; ADR-0026)"
      responses: { "200": { description: "{auto_ids}" }, "503": { $ref: "#/components/responses/Error" } }
    put:
      summary: "Сменить режим авто-ID ({auto_ids}): применяется сразу и сохраняется в config.yaml; неизвестный режим — 400 api.auto_ids_bad"
      responses: { "200": { description: "{auto_ids}" }, "400": { $ref: "#/components/responses/Error" } }
...
  /api/v1/projects/import:
    post:
      summary: "Импортировать проект ({name, content}); новый проект выключен (SEC-7)"
      responses: { "200": { description: "{id}" }, "400": { $ref: "#/components/responses/Error" } }
  /api/v1/projects/{id}/{action}:
    post:
      summary: Включить или выключить проект (action = enable | disable)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: action, in: path, required: true, schema: { enum: [enable, disable] } }
      responses: { "200": { description: Выполнено }, "400": { $ref: "#/components/responses/Error" } }
...
  /api/v1/events/{project}/{event}/{action}:
    post:
      summary: Запустить (run, ждёт завершения), включить (enable) или выключить (disable) событие
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: event, in: path, required: true, schema: { type: string } }
        - { name: action, in: path, required: true, schema: { enum: [run, enable, disable] } }
      responses: { "200": { description: Выполнено }, "404": { $ref: "#/components/responses/Error" } }
  /api/v1/vars/{project}:
    get:
      summary: Переменные активного проекта
      parameters: [ { name: project, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: "{vars: {...}}" }, "404": { $ref: "#/components/responses/Error" } }
  /api/v1/vars/{project}/{name}:
    put:
      summary: "Задать переменную ({value})"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Выполнено } }
  /api/v1/wait/key:
    post:
      summary: "Ждать нажатия клавиши ({key, timeout_ms}); 408 — время вышло"
      responses: { "200": { description: Нажата }, "408": { $ref: "#/components/responses/Error" } }
...
  /api/v1/settings/interface:
    get:
      summary: "Язык и тема окна из config.yaml: {language: ru|en|\"\", theme: system|light|dark}"
      responses: { "200": { description: Язык и тема } }
    put:
      summary: "Сохранить язык (language в начале config.yaml) и тему (modules.api.theme); неверные — 400 api.interface_bad"
      responses:
        "200": { description: Сохранённые значения }
        "400": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: mkey_session }
  responses:
    Error:
      description: Ошибка в едином формате
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: dsl.unknown_key_hint }
                  message: { type: string, example: "неизвестная клавиша «Mous0» — возможно, вы имели в виду «Mouse0»?" }
                  details:
                    type: object
                    properties:
                      pos:
                        type: object
                        properties:
                          line: { type: integer }
                          col: { type: integer }
                      args: { type: object, additionalProperties: { type: string } }
  schemas:
    TextRequest:
      type: object
      required: [text]
      additionalProperties: false
      properties:
        text: { type: string }
    InputStatus:
      type: object
      properties:
        open: { type: integer }
        denied: { type: array, items: { type: string } }
...
    Node:
      type: object
      description: Узел дерева разбора (internal/lib/dsl.Node); используются поля, относящиеся к type.
      properties:
        type: { type: string, enum: [tap, down, up, up_all, axis, pause, text, command, group] }
        pos: { type: object, properties: { line: { type: integer }, col: { type: integer } } }
        keys:
          type: array
          items:
            type: object
            properties:
              device: { type: string }
              name: { type: string }
              code: { type: integer }
        repeat: { type: integer }
        hold_ms: { type: integer }
        value: { type: number }
        min_ms: { type: integer }
        max_ms: { type: integer }
        text: { type: string }
        command: { type: string }
        args: { type: array, items: { type: object } }
        children: { type: array, items: { $ref: "#/components/schemas/Node" } }